requirements.txt
click==6.7
cycler==0.10.0
funcsigs==1.0.2
functools32==3.2.3.post2
h5py==2.7.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
matplotlib==2.0.0
mock==2.0.0
numpy==1.12.1
olefile==0.44
opencv-python==3.2.0.6
pbr==2.0.0
Pillow==4.0.0
protobuf==3.2.0
psutil==5.2.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
subprocess32==3.2.7
tensorflow==1.0.0
Werkzeug==0.12.1
pykafka
psycopg2
confluent-kafka
